Modifies the CI.Core script to attempt to load the Azure account for provisioning and deploying Azure services.
The credentials can be set via Environment Variables if the user has not already saved credentials to the local machine.  This feature is mainly to support continuous integration systems where the Azure credentials need to be stored externally and set on every build.

The Environment variables to set are:
AzureDeployUser
AzureDeployPass

This user cannot be a normal Microsoft Live account (even though they work when prompted and entered manually, see https://azure.microsoft.com/en-us/blog/azure-automation-authenticating-to-azure-using-azure-active-directory/).  This user must exist in an Active Directory either hosted in Azure or elsewhere.